+++ NEWS.md
@@ -3,6 +3,7 @@
 * `.edma_fit_np` is more preformant (original implementation
   retained as `.edma_fit_np_old` for comparison);
   it also exposes mean/variance for distances for future uses.
+* `edma_simulate_data` is also much faster now.
 
 # Version 0.1-3, 2020-07-05